Async PDFium rendering, safety and UI polish
Make PDF operations asynchronous and move heavy work to a blocking pool (run_pdfium). Add a process-wide PDFIUM_GATE and shared Pdfium instance to serialize access and safely drop library state. Harden rendering (dimension/pixel checks, BGRx→RGBA conversion) with tests and switch pdfium-render features. Convert many Tauri commands to async, improve form write queuing and undo handling. Major frontend polish: new Icon component, revamped TabBar and Viewer (render scheduling, DPR cap, zoom anchoring, accessibility), CSS/theme updates, added app icons, and updated tauri capabilities and build scripts.

+++ crates/pdf-core/src/forms.rs
@@ -196,7 +196,7 @@ impl Document {
             let pages = doc.pages();
             let page_count = pages.len();
             for pi in 0..page_count {
-                let page = match pages.get(pi as u16) {
+                let page = match pages.get(pi) {
                     Ok(p) => p,
                     Err(_) => continue,
                 };
